async-api/requirements/api.txt
aioredis==1.3.1
aiocache==0.11.1
fastapi==0.70.0
orjson==3.6.4
uvicorn==0.15.0
uvloop==0.16.0
backoff==1.11.1
jsonpickle==2.0.0
elasticsearch_dsl==7.4.0
ujson==5.0.0
msgpack==1.0.3
aioredis==1.3.1
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-fastapi==0.29b0
python-jose[cryptography]==3.3.0